+++ src/main/clojure/aerospike_clj/client.clj
@@ -3,6 +3,7 @@
   (:require [clojure.string :as s]
             [clojure.tools.logging :as log]
             [promesa.core :as p]
+            [promesa.exec :as p-exec]
             [aerospike-clj.policy :as policy]
             [aerospike-clj.bins :as bins]
             [aerospike-clj.utils :as utils]
@@ -11,18 +12,20 @@
             [aerospike-clj.listeners]
             [aerospike-clj.aerospike-record :as record]
             [aerospike-clj.protocols :as pt])
-  (:import [java.time Instant]
-           [java.util List Collection ArrayList]
-           [com.aerospike.client AerospikeClient Key Bin Operation BatchRead]
-           [com.aerospike.client.async EventLoop NioEventLoops EventLoops]
-           [com.aerospike.client.cluster Node]
-           [com.aerospike.client.policy Policy BatchPolicy ClientPolicy
+  (:import (java.time Instant)
+           (java.util List Collection ArrayList Arrays)
+           (com.aerospike.client AerospikeClient Key Bin Operation BatchRead)
+           (com.aerospike.client.async EventLoop NioEventLoops EventLoops)
+           (com.aerospike.client.cluster Node)
+           (com.aerospike.client.policy Policy BatchPolicy ClientPolicy
                                         RecordExistsAction WritePolicy ScanPolicy
-                                        InfoPolicy]
-           [com.aerospike.client Key Host]
-           [aerospike_clj.listeners AsyncExistsListener AsyncDeleteListener AsyncWriteListener
+                                        InfoPolicy)
+           (com.aerospike.client Key Host BatchRecord)
+           (aerospike_clj.listeners AsyncExistsListener AsyncDeleteListener AsyncWriteListener
                                     AsyncInfoListener AsyncRecordListener AsyncRecordSequenceListener
-                                    AsyncBatchListListener AsyncExistsArrayListener]))
+                                    AsyncBatchListListener AsyncExistsArrayListener AsyncBatchOperateListListener)
+           (com.aerospike.client.listener BatchOperateListListener)
+           (java.util.concurrent Executor)))
 
 (def
   ^{:doc   "The 0 date reference for returned record TTL"
@@ -62,10 +65,11 @@
         (p/catch (fn [op-exception]
                    (pt/on-failure client-events op-name op-exception index op-start-time))))))
 
-(defn- register-events [op-future client-events op-name index op-start-time]
-  (if (empty? client-events)
-    op-future
-    (reduce (client-events-reducer op-name index op-start-time) op-future client-events)))
+(defn- register-events [op-future default-client-events op-name index op-start-time conf]
+  (let [client-events (:client-events conf default-client-events)]
+    (if (empty? client-events)
+      op-future
+      (reduce (client-events-reducer op-name index op-start-time) op-future client-events))))
 
 (extend-protocol pt/UserKey
   Key
@@ -75,11 +79,12 @@
   (create-key ^Key [this as-namespace set-name]
     (as-key/create-key this as-namespace set-name)))
 
-(defn- batch-read->map [^BatchRead batch-read]
-  (let [k (.key batch-read)]
-    (-> (record/record->map (.record batch-read))
+(defn- batch-record->map [^BatchRecord batch-record]
+  (let [k (.key batch-record)]
+    (-> (record/record->map (.record batch-record))
         (assoc :index (.toString (.userKey k)))
-        (assoc :set (.setName k)))))
+        (assoc :set (.setName k))
+        (assoc :result-code (.resultCode batch-record)))))
 
 (defn- map->batch-read ^BatchRead [batch-read-map dbns]
   (let [k ^Key (pt/create-key (:index batch-read-map) dbns (:set batch-read-map))]
@@ -89,7 +94,7 @@
       (BatchRead. k ^"[Ljava.lang.String;" (utils/v->array String (:bins batch-read-map))))))
 
 ;; put
-(defn- put* [^AerospikeClient client ^EventLoops event-loops dbns client-events index data policy set-name]
+(defn- put* [^AerospikeClient client ^EventLoops event-loops dbns client-events index data policy set-name conf]
   (let [bins       (bins/data->bins data)
         op-future  (p/deferred)
         start-time (System/nanoTime)]
@@ -99,10 +104,11 @@
           ^WritePolicy policy
           ^Key (pt/create-key index dbns set-name)
           ^"[Lcom.aerospike.client.Bin;" bins)
-    (register-events op-future client-events :write index start-time)))
+    (register-events op-future client-events :write index start-time conf)))
 
 (deftype SimpleAerospikeClient [client
                                 el
+                                ^Executor completion-executor
                                 hosts
                                 dbns
                                 client-events
@@ -133,10 +139,14 @@
               ^Policy (:policy conf)
               ^Key (pt/create-key index dbns set-name)
               ^"[Ljava.lang.String;" (utils/v->array String bin-names)))
-      (let [p (p/chain op-future
-                       record/record->map
-                       (:transcoder conf identity))]
-        (register-events p client-events :read index start-time))))
+      (-> op-future
+          (p/then' (fn [res]
+                     (let [transcoder (:transcoder conf identity)]
+                       (-> res
+                           record/record->map
+                           transcoder)))
+                   completion-executor)
+          (register-events client-events :read index start-time conf))))
 
   (get-single-no-meta [this index set-name]
     (pt/get-single this index set-name {:transcoder :payload}))
@@ -155,7 +165,9 @@
                (AsyncExistsListener. op-future)
                ^Policy (:policy conf)
                ^Key (pt/create-key index dbns set-name))
-      (register-events op-future client-events :exists index start-time)))
+      (-> op-future
+          (p/then' identity completion-executor)
+          (register-events client-events :exists index start-time conf))))
 
   (get-batch [this batch-reads]
     (pt/get-batch this batch-reads {}))
@@ -169,10 +181,10 @@
             (AsyncBatchListListener. op-future)
             ^BatchPolicy (:policy conf)
             ^List batch-reads-arr)
-      (let [d (p/chain op-future
-                       #(mapv batch-read->map %)
-                       (:transcoder conf identity))]
-        (register-events d client-events :read-batch nil start-time))))
+      (-> op-future
+          (p/then' #(mapv batch-record->map %) completion-executor)
+          (p/then' (:transcoder conf identity))
+          (register-events client-events :read-batch nil start-time conf))))
 
   (exists-batch [this indices]
     (pt/exists-batch this indices {}))
@@ -180,16 +192,16 @@
   (exists-batch [_this indices conf]
     (let [op-future  (p/deferred)
           start-time (System/nanoTime)
+          transcoder (:transcoder conf identity)
           indices    (utils/v->array Key (mapv #(pt/create-key (:index %) dbns (:set %)) indices))]
       (.exists ^AerospikeClient client
                ^EventLoop (.next ^EventLoops el)
                (AsyncExistsArrayListener. op-future)
                ^BatchPolicy (:policy conf)
                ^"[Lcom.aerospike.client.Key;" indices)
-      (let [d (p/chain op-future
-                       vec
-                       (:transcoder conf identity))]
-        (register-events d client-events :exists-batch nil start-time))))
+      (-> op-future
+          (p/then' (comp transcoder vec) completion-executor)
+          (register-events client-events :exists-batch nil start-time conf))))
 
   pt/AerospikeWriteOps
   (put [this index set-name data expiration]
@@ -203,7 +215,8 @@
           index
           ((:transcoder conf identity) data)
           (:policy conf (policy/write-policy client expiration))
-          set-name))
+          set-name
+          conf))
 
   (create [this index set-name data expiration]
     (pt/create this index set-name data expiration {}))
@@ -216,7 +229,8 @@
           index
           ((:transcoder conf identity) data)
           (policy/create-only-policy client expiration)
-          set-name))
+          set-name
+          conf))
 
   (put-multiple [this indices set-names payloads expirations]
     (pt/put-multiple this indices set-names payloads expirations {}))
@@ -239,7 +253,8 @@
           index
           ((:transcoder conf identity) data)
           (policy/set-policy client expiration)
-          set-name))
+          set-name
+          conf))
 
   (replace-only [this index set-name data expiration]
     (pt/replace-only this index set-name data expiration {}))
@@ -252,7 +267,8 @@
           index
           ((:transcoder conf identity) data)
           (policy/replace-only-policy client expiration)
-          set-name))
+          set-name
+          conf))
 
   (update [this index set-name new-record generation new-expiration]
     (pt/update this index set-name new-record generation new-expiration {}))
@@ -265,7 +281,8 @@
           index
           ((:transcoder conf identity) new-record)
           (policy/update-policy client generation new-expiration)
-          set-name))
+          set-name
+          conf))
 
   (add-bins [this index set-name new-data new-expiration]
     (pt/add-bins this index set-name new-data new-expiration {}))
@@ -278,9 +295,13 @@
           index
           ((:transcoder conf identity) new-data)
           (policy/update-only-policy client new-expiration)
-          set-name))
+          set-name
+          conf))
 
-  (touch [_this index set-name expiration]
+  (touch [this index set-name expiration]
+    (pt/touch this index set-name expiration {}))
+
+  (touch [_this index set-name expiration conf]
     (let [op-future  (p/deferred)
           start-time (System/nanoTime)]
       (.touch ^AerospikeClient client
@@ -288,7 +309,9 @@
               (AsyncWriteListener. op-future)
               ^WritePolicy (policy/write-policy client expiration RecordExistsAction/UPDATE_ONLY)
               ^Key (pt/create-key index dbns set-name))
-      (register-events op-future client-events :touch index start-time)))
+      (-> op-future
+          (p/then' identity completion-executor)
+          (register-events client-events :touch index start-time conf))))
 
   pt/AerospikeDeleteOps
   (delete [this index set-name]
@@ -302,7 +325,9 @@
                (AsyncDeleteListener. op-future)
                ^WritePolicy (:policy conf)
                ^Key (pt/create-key index dbns set-name))
-      (register-events op-future client-events :delete index start-time)))
+      (-> op-future
+          (p/then' identity completion-executor)
+          (register-events client-events :delete index start-time conf))))
 
   (delete-bins [this index set-name bin-names new-expiration]
     (pt/delete-bins this index set-name bin-names new-expiration {}))
@@ -318,7 +343,9 @@
             ^WritePolicy policy
             ^Key (pt/create-key index dbns set-name)
             ^"[Lcom.aerospike.client.Bin;" (utils/v->array Bin (mapv bins/set-bin-as-null bin-names)))
-      (register-events op-future client-events :write index start-time)))
+      (-> op-future
+          (p/then' identity completion-executor)
+          (register-events client-events :write index start-time conf))))
 
   pt/AerospikeSingleIndexBatchOps
   (operate [this index set-name expiration operations]
@@ -335,7 +362,33 @@
                   ^WritePolicy (:policy conf (policy/write-policy client expiration RecordExistsAction/UPDATE))
                   ^Key (pt/create-key index dbns set-name)
                   (utils/v->array Operation operations))
-        (register-events (p/then op-future record/record->map) client-events :operate index start-time))))
+        (-> op-future
+            (p/then' record/record->map completion-executor)
+            (register-events client-events :operate index start-time conf)))))
+
+  pt/AerospikeBatchOps
+  (batch-operate [this batch-records]
+    (pt/batch-operate this batch-records {}))
+
+  (batch-operate [_this batch-records conf]
+    (let [op-future  (p/deferred)
+          policy     (:policy conf)
+          batch-list (if (list? batch-records)
+                       batch-records
+                       (->> batch-records
+                            (utils/v->array BatchRecord)
+                            (Arrays/asList)))
+          start-time (System/nanoTime)
+          transcoder (:transcoder conf identity)]
+      (.operate ^AerospikeClient client
+                ^EventLoop (.next ^EventLoops el)
+                ^BatchOperateListListener (AsyncBatchOperateListListener. op-future)
+                ^BatchPolicy policy
+                ^List batch-list)
+      (-> op-future
+          (p/then' (comp transcoder #(mapv batch-record->map %)) completion-executor)
+          (register-events client-events :batch-operate nil start-time conf))))
+
 
   pt/AerospikeSetOps
   (scan-set [_this aero-namespace set-name conf]
@@ -351,7 +404,9 @@
                 aero-namespace
                 set-name
                 (when bin-names ^"[Ljava.lang.String;" (utils/v->array String bin-names)))
-      (register-events op-future client-events :scan nil start-time)))
+      (-> op-future
+          (p/then' identity completion-executor)
+          (register-events client-events :scan nil start-time conf))))
 
   pt/AerospikeAdminOps
   (info [this node info-commands]
@@ -366,7 +421,9 @@
              ^InfoPolicy (:policy conf (.infoPolicyDefault ^AerospikeClient client))
              ^Node node
              (into-array String info-commands))
-      (register-events op-future client-events :info nil start-time)))
+      (-> op-future
+          (p/then' identity completion-executor)
+          (register-events client-events :info nil start-time conf))))
 
   (get-nodes [_this]
     (into [] (.getNodes ^AerospikeClient client)))
@@ -432,12 +489,14 @@
   ([hosts aero-ns]
    (init-simple-aerospike-client hosts aero-ns {}))
   ([hosts aero-ns conf]
-   (let [close-event-loops? (nil? (:event-loops conf))
-         event-loops        (or (:event-loops conf) (create-event-loops conf))
-         client-policy      (:client-policy conf (policy/create-client-policy event-loops conf))]
+   (let [close-event-loops?  (nil? (:event-loops conf))
+         event-loops         (or (:event-loops conf) (create-event-loops conf))
+         completion-executor (:completion-executor conf p-exec/default-executor)
+         client-policy       (:client-policy conf (policy/create-client-policy event-loops conf))]
      (log/info (format "Starting aerospike client for hosts %s with username %s" hosts (get conf "username")))
      (->SimpleAerospikeClient (create-client hosts client-policy (:port conf 3000))
                               event-loops
+                              completion-executor
                               hosts
                               aero-ns
                               (utils/vectorize (:client-events conf))