+++ src/main/clojure/aerospike_clj/listeners.clj
@@ -1,10 +1,10 @@
 (ns aerospike-clj.listeners
   (:require [promesa.core :as p]
             [aerospike-clj.aerospike-record :as record])
-  (:import [java.util List Map]
-           [com.aerospike.client Key Record AerospikeException AerospikeException$QueryTerminated]
-           [com.aerospike.client.listener RecordListener WriteListener DeleteListener
-            ExistsListener BatchListListener RecordSequenceListener InfoListener ExistsArrayListener]))
+  (:import (java.util List Map)
+           (com.aerospike.client Key Record AerospikeException AerospikeException$QueryTerminated)
+           (com.aerospike.client.listener RecordListener WriteListener DeleteListener
+                                          ExistsListener BatchListListener RecordSequenceListener InfoListener ExistsArrayListener BatchOperateListListener)))
 
 (deftype AsyncExistsListener [op-future]
   ExistsListener
@@ -66,3 +66,10 @@
     (p/reject! op-future ex))
   (^void onSuccess [_this ^"[Lcom.aerospike.client.Key;" _keys ^"[Z" exists]
     (p/resolve! op-future exists)))
+
+(deftype AsyncBatchOperateListListener [op-future]
+  BatchOperateListListener
+  (^void onSuccess [_this ^List records ^boolean _status]
+    (p/resolve! op-future records))
+  (^void onFailure [_this ^AerospikeException ex]
+    (p/reject! op-future ex)))