+++ src/main/clojure/aerospike_clj/policy.clj
@@ -1,10 +1,10 @@
 (ns aerospike-clj.policy
   (:import [com.aerospike.client AerospikeClient]
            [com.aerospike.client.async EventPolicy]
-           #_{:clj-kondo/ignore [:unused-import]}
+    #_{:clj-kondo/ignore [:unused-import]}
            [com.aerospike.client.policy Policy ClientPolicy WritePolicy RecordExistsAction
                                         GenerationPolicy BatchPolicy CommitLevel
-                                        AuthMode ReadModeAP ReadModeSC Priority Replica]))
+                                        AuthMode ReadModeAP ReadModeSC Replica BatchWritePolicy]))
 
 (defmacro set-java [obj conf obj-name]
   `(when (some? (get ~conf ~obj-name))
@@ -30,7 +30,6 @@
     (set-java-enum p conf "ReadModeAP")
     (set-java-enum p conf "ReadModeSC")
     (set-java p conf "maxRetries")
-    (set-java-enum p conf "Priority")
     (set-java-enum p conf "Replica")
     (set-java p conf "sendKey")
     (set-java p conf "sleepBetweenRetries")
@@ -39,13 +38,29 @@
     (set-java p conf "totalTimeout")
     p))
 
+(defn map->batch-write-policy
+  "Create a `BatchWritePolicy` from a map. Enumeration names should start with capitalized letter.
+  This function is slow due to possible reflection."
+  ^BatchWritePolicy [conf]
+  (let [p (BatchWritePolicy.)]
+    (set-java-enum p conf "RecordExistsAction")
+    (set-java-enum p conf "CommitLevel")
+    (set-java-enum p conf "GenerationPolicy")
+    (set-java p conf "filterExp")
+    (set-java p conf "generation")
+    (set-java p conf "expiration")
+    (set-java p conf "durableDelete")
+    (set-java p conf "sendKey")
+    p))
+
 (defn map->batch-policy
-  "Create a (read) `BatchPolicy` from a map.
+  "Create a `BatchPolicy` from a map.
   This function is slow due to possible reflection."
   ^BatchPolicy [conf]
-  (let [bp   (BatchPolicy.)
+  (let [bp   (BatchPolicy. (map->policy conf))
         conf (merge {"timeoutDelay" 3000} conf)]
     (set-java bp conf "allowInline")
+    (set-java bp conf "respondAllKeys")
     (set-java bp conf "maxConcurrentThreads")
     (set-java bp conf "sendSetName")
     bp))
@@ -73,6 +88,18 @@
    (write-policy client expiration (RecordExistsAction/REPLACE)))
   (^WritePolicy [client expiration record-exists-action]
    (let [wp (WritePolicy. (.getWritePolicyDefault ^AerospikeClient client))]
+     (set! (.expiration wp) expiration)
+     (set! (.recordExistsAction wp) record-exists-action)
+     wp)))
+
+(defn batch-write-policy
+  "Create a write policy to be passed to put methods via `{:policy wp}`.
+  Also used in `update` and `create`.
+  The default policy in case the record exists is `RecordExistsAction/UPDATE`."
+  (^BatchWritePolicy [client expiration]
+   (batch-write-policy client expiration (RecordExistsAction/UPDATE)))
+  (^BatchWritePolicy [client expiration record-exists-action]
+   (let [wp (BatchWritePolicy. (.getBatchWritePolicyDefault ^AerospikeClient client))]
      (set! (.expiration wp) expiration)
      (set! (.recordExistsAction wp) record-exists-action)
      wp)))
@@ -160,6 +187,9 @@
     (set! (.readPolicyDefault cp) (get conf "readPolicyDefault" (map->policy conf)))
     (set! (.writePolicyDefault cp) (get conf "writePolicyDefault" (map->write-policy conf)))
     (set! (.batchPolicyDefault cp) (get conf "batchPolicyDefault" (map->batch-policy conf)))
+    (set! (.batchParentPolicyWriteDefault cp) (get conf "batchParentPolicyWriteDefault" (map->batch-policy conf)))
+    (set! (.batchWritePolicyDefault cp) (get conf "batchWritePolicyDefault" (map->batch-write-policy conf)))
+
     (set-java-enum cp conf "AuthMode")
     (set-java cp conf "clusterName")
     (set-java cp conf "connPoolsPerNode")