+++ test/aerospike_clj/integration/integration_test.clj
@@ -1,53 +1,58 @@
 (ns ^{:author      "Ido Barkan"
-      :integration true
-      :doc         "Integration tests. Requires a local Aerospike instance.
-                   To run instances locally inside docker containers:
-                   $ docker run -d --name aerospike -p 3000:3000 -p 3001:3001 -p 3002:3002 -p 3003:3003 aerospike:4.9.0.11"}
-  aerospike-clj.integration-test
+      :integration true}
+  aerospike-clj.integration.integration-test
   (:require [clojure.test :refer [deftest testing is use-fixtures]]
+            [cheshire.core :as json]
+            [aerospike-clj.integration.aerospike-setup :as as-setup]
             [aerospike-clj.client :as client]
             [aerospike-clj.protocols :as pt]
             [aerospike-clj.policy :as policy]
             [aerospike-clj.key :as as-key]
-            [cheshire.core :as json])
-  (:import [com.aerospike.client Value AerospikeClient]
-           [com.aerospike.client.cdt ListOperation ListPolicy ListOrder ListWriteFlags ListReturnType
-                                     MapOperation MapPolicy MapOrder MapWriteFlags MapReturnType CTX]
-           [com.aerospike.client.policy Priority ReadModeSC ReadModeAP Replica GenerationPolicy RecordExistsAction
-                                        WritePolicy BatchPolicy Policy]
-           [java.util HashMap ArrayList]
-           [java.util.concurrent ExecutionException]
-           [clojure.lang PersistentArrayMap]
-           [aerospike_clj.client SimpleAerospikeClient]))
+            [aerospike-clj.utils :as utils]
+            [spy.core :as spy])
+  (:import (com.aerospike.client Value AerospikeClient BatchWrite Operation Bin)
+           (com.aerospike.client.cdt ListOperation ListPolicy ListOrder ListWriteFlags ListReturnType
+                                     MapOperation MapPolicy MapOrder MapWriteFlags MapReturnType CTX)
+           (com.aerospike.client.policy ReadModeSC ReadModeAP Replica GenerationPolicy RecordExistsAction
+                                        WritePolicy BatchPolicy Policy CommitLevel BatchWritePolicy)
+           (java.util HashMap ArrayList)
+           (java.util.concurrent ExecutionException)
+           (clojure.lang PersistentArrayMap)
+           (aerospike_clj.client SimpleAerospikeClient)
+           (com.aerospike.client.exp Exp)))
 
 (def _set "set")
 (def _set2 "set2")
 (def as-namespace "test")
 (def ^:dynamic *c* nil)
+(def ^:dynamic *as-hosts* nil)
 (def TTL 5)
 
-(defn db-connection [test-fn]
-  (binding [*c* (client/init-simple-aerospike-client
-                  ["localhost"]
-                  as-namespace)]
-    (test-fn)
-    (pt/stop *c*)))
 
-(use-fixtures :once db-connection)
+(defn with-db-connection [test-fn]
+  (let [as-hosts [(as-setup/get-host-and-port)]]
+    (binding [*c*        (client/init-simple-aerospike-client
+                           as-hosts
+                           as-namespace)
+              *as-hosts* as-hosts]
+      (test-fn)
+      (pt/stop *c*))))
+
+(use-fixtures :once as-setup/with-aerospike with-db-connection)
 
 (deftest client-creation
-  (let [c (client/init-simple-aerospike-client ["localhost"] "test")]
+  (let [c (client/init-simple-aerospike-client *as-hosts* as-namespace)]
     (is c)
-    (is (= ["localhost"] (.-hosts ^SimpleAerospikeClient c))))
+    (is (= *as-hosts* (.-hosts ^SimpleAerospikeClient c))))
 
   (letfn [(no-password? [ex]
             (let [conf (:conf (ex-data ex))]
               (and conf (not (contains? conf "password")))))]
-    (let [ex (is (thrown-with-msg? Exception #"unbounded delay queue" (client/init-simple-aerospike-client ["localhost"] "test" {"maxCommandsInProcess" 1})))]
+    (let [ex (is (thrown-with-msg? Exception #"unbounded delay queue" (client/init-simple-aerospike-client *as-hosts* as-namespace {"maxCommandsInProcess" 1})))]
       (is (no-password? ex)))
 
     (with-redefs [client/create-event-loops (constantly nil)]
-      (let [ex (is (thrown-with-msg? Exception #"event-loops" (client/init-simple-aerospike-client ["localhost"] "test")))]
+      (let [ex (is (thrown-with-msg? Exception #"event-loops" (client/init-simple-aerospike-client *as-hosts* as-namespace)))]
         (is (no-password? ex))))))
 
 (deftest health
@@ -141,7 +146,7 @@
     (is (true? @(pt/create *c* k _set data TTL)))
     (testing "clojure maps can be serialized as-is"
       (let [v @(pt/get-single-no-meta *c* k _set)]
-        (is (= data v)) ;; per value it is identical
+        (is (= data v))                                     ;; per value it is identical
         (is (= PersistentArrayMap (type v)))))))
 
 (deftest put-multiple-bins-get-clj-map
@@ -153,11 +158,11 @@
     (is (true? @(pt/create *c* k _set data TTL)))
     (testing "clojure maps can be serialized from bins"
       (let [v @(pt/get-single-no-meta *c* k _set)]
-        (is (= (get data "foo") (get v "foo"))) ;; per value it is identical
-        (is (= (get data "bar") (get v "bar"))) ;; true value returns the same after being sanitized/desanitized
-        (is (= (get data "baz") (get v "baz"))) ;; false value returns the same after being sanitized/desanitized
-        (is (= (get data "qux") (get v "qux"))) ;; nil value retuns the same after being sanitized/desanitized
-        (is (= PersistentArrayMap (type v))) ;; converted back to a Clojure map instead of HashMap
+        (is (= (get data "foo") (get v "foo")))             ;; per value it is identical
+        (is (= (get data "bar") (get v "bar")))             ;; true value returns the same after being sanitized/desanitized
+        (is (= (get data "baz") (get v "baz")))             ;; false value returns the same after being sanitized/desanitized
+        (is (= (get data "qux") (get v "qux")))             ;; nil value retuns the same after being sanitized/desanitized
+        (is (= PersistentArrayMap (type v)))                ;; converted back to a Clojure map instead of HashMap
         (is (true? (map? v)))))))
 
 (deftest get-single-multiple-bins
@@ -170,7 +175,7 @@
       (let [v1 @(pt/get-single *c* k _set {} ["foo"])
             v2 @(pt/get-single *c* k _set {} ["bar"])
             v3 @(pt/get-single *c* k _set {} ["baz"])
-            v4 @(pt/get-single *c* k _set {})] ;; getting all bins for the record
+            v4 @(pt/get-single *c* k _set {})]              ;; getting all bins for the record
         (is (= (get data "foo") (get (:payload v1) "foo")))
         (is (= (get data "bar") (get (:payload v2) "bar")))
         (is (= (get data "baz") (get (:payload v3) "baz")))
@@ -200,7 +205,7 @@
         new-data {"qux" [(rand-int 1000)]}
         k        (random-key)]
     (is (true? @(pt/create *c* k _set data TTL)))
-    (is (true? @(pt/add-bins *c* k _set new-data TTL))) ;; adding value to bin
+    (is (true? @(pt/add-bins *c* k _set new-data TTL)))     ;; adding value to bin
     (testing "bin values can be added to existing records"
       (let [v @(pt/get-single-no-meta *c* k _set)]
         (is (= v (merge data new-data)))
@@ -219,7 +224,7 @@
         bin-keys ["foo" "bar" "baz"]
         k        (random-key)]
     (is (true? @(pt/create *c* k _set data TTL)))
-    (is (true? @(pt/delete-bins *c* k _set bin-keys TTL))) ;; removing value from bin
+    (is (true? @(pt/delete-bins *c* k _set bin-keys TTL)))  ;; removing value from bin
     (testing "bin values can be removed from existing records"
       (let [v @(pt/get-single-no-meta *c* k _set)]
         (is (= v (apply dissoc data bin-keys)))
@@ -519,26 +524,62 @@
       (is (nil? (first (:payload @(set-pop "foo")))))
       (is (= #{"bar"} (set-getall))))))
 
+(deftest batch-operate
+  (testing "mixed cdt operations on multiple keys"
+    (letfn [(create-batch-write-record [list-bin map-bin map-key ^String string-bin k v]
+              (let [as-key (pt/create-key k as-namespace _set)]
+                (BatchWrite. as-key (utils/v->array Operation [(ListOperation/append list-bin (Value/get (str v)) nil)
+                                                               (MapOperation/put (MapPolicy.) map-bin (Value/get map-key) (Value/get (str v)) nil)
+                                                               (Operation/put (Bin. string-bin (str v)))]))))
+            (create-touch-record [k _v]
+              (let [as-key (pt/create-key k as-namespace _set)]
+                (BatchWrite. as-key (utils/v->array Operation [(Operation/touch)]))))
+            (create-read-batch-record [k]
+              {:index k :set _set})]
+      (let [list-bin                      "list"
+            map-bin                       "map"
+            map-key                       "test-key"
+            string-bin                    "string"
+            ks                            (take 3 (repeatedly random-key))
+            expected-write-result-payload {:result-code 0
+                                           :payload     {list-bin   1
+                                                         map-bin    1
+                                                         string-bin nil}}
+            expected-read-payloads        (mapv (fn [^String val]
+                                                  (hash-map map-bin {map-key (str val)}
+                                                            list-bin [(str val)]
+                                                            string-bin (str val))) (range 3))
+            batch-write-records           (mapv (partial create-batch-write-record
+                                                         list-bin
+                                                         map-bin
+                                                         map-key
+                                                         string-bin) ks (range))
+            batch-touch-records           (mapv create-touch-record ks (range))
+            batch-read-records            (mapv create-read-batch-record ks)]
+        (is (every? #(= expected-write-result-payload (select-keys % [:result-code :payload]))
+                    @(pt/batch-operate *c* batch-write-records)))
+        (is (= expected-read-payloads
+               (mapv :payload @(pt/get-batch *c* batch-read-records))))
+        (is (every? #(zero? (:result-code %)) @(pt/batch-operate *c* batch-touch-records)))))))
+
 (deftest default-read-policy
   (let [rp (.getReadPolicyDefault ^AerospikeClient (.-client ^SimpleAerospikeClient *c*))]
-    (is (= Priority/DEFAULT (.priority rp))) ;; Priority of request relative to other transactions. Currently, only used for scans.
-    (is (= ReadModeAP/ONE (.readModeAP rp))) ;; Involve single node in the read operation.
-    (is (= Replica/SEQUENCE (.replica rp))) ;; Try node containing master partition first.
+    (is (= Replica/SEQUENCE (.replica rp)))                 ;; Try node containing master partition first.
     ;; If connection fails, all commands try nodes containing replicated partitions.
     ;; If socketTimeout is reached, reads also try nodes containing replicated partitions,
     ;; but writes remain on master node.)))
     ;; This option requires ClientPolicy.requestProleReplicas to be enabled in order to function properly.
-    (is (= 30000 (.socketTimeout rp))) ;; 30 seconds default
-    (is (zero? (.totalTimeout rp))) ;; no time limit
-    (is (= 3000 (.timeoutDelay rp))) ;; no delay, connection closed on timeout
-    (is (= 2 (.maxRetries rp))) ;; initial attempt + 2 retries = 3 attempts
-    (is (zero? (.sleepBetweenRetries rp))) ;; do not sleep between retries
-    (is (false? (.sendKey rp))) ;; do not send the user defined key
-    (is (= ReadModeSC/SESSION (.readModeSC rp))))) ;; Ensures this client will only see an increasing sequence of record versions. Server only reads from master. This is the default..
+    (is (= 30000 (.socketTimeout rp)))                      ;; 30 seconds default
+    (is (= 1000 (.totalTimeout rp)))                        ;; total timeout of 1 second
+    (is (= 3000 (.timeoutDelay rp)))                        ;; no delay, connection closed on timeout
+    (is (= 2 (.maxRetries rp)))                             ;; initial attempt + 2 retries = 3 attempts
+    (is (zero? (.sleepBetweenRetries rp)))                  ;; do not sleep between retries
+    (is (false? (.sendKey rp)))                             ;; do not send the user defined key
+    (is (= ReadModeSC/SESSION (.readModeSC rp)))))          ;; Ensures this client will only see an increasing sequence of record versions. Server only reads from master. This is the default.
 
 (deftest configure-read-and-batch-policy
   (let [c  (client/init-simple-aerospike-client
-             ["localhost"] "test"
+             *as-hosts* as-namespace
              {"readPolicyDefault"  (policy/map->policy {"ReadModeAP"          "ALL"
                                                         "ReadModeSC"          "LINEARIZE"
                                                         "maxRetries"          1
@@ -555,7 +596,6 @@
 
         rp ^Policy (.getReadPolicyDefault ^AerospikeClient (.-client ^SimpleAerospikeClient c))
         bp ^BatchPolicy (.getBatchPolicyDefault ^AerospikeClient (.-client ^SimpleAerospikeClient c))]
-    (is (= Priority/DEFAULT (.priority rp)))
     (is (= ReadModeAP/ALL (.readModeAP rp)))
     (is (= Replica/RANDOM (.replica rp)))
     (is (= 1000 (.socketTimeout rp)))
@@ -572,23 +612,24 @@
 
 (deftest default-write-policy
   (let [rp ^WritePolicy (.getWritePolicyDefault ^AerospikeClient (.-client ^SimpleAerospikeClient *c*))]
-    (is (= ReadModeAP/ONE (.readModeAP rp))) ;; Involve master only in the read operation.
-    (is (= Replica/SEQUENCE (.replica rp))) ;; Try node containing master partition first.
+    (is (= ReadModeAP/ONE (.readModeAP rp)))                ;; Involve master only in the read operation.
+    (is (= Replica/SEQUENCE (.replica rp)))                 ;; Try node containing master partition first.
     ;; If connection fails, all commands try nodes containing replicated partitions.
     ;; If socketTimeout is reached, reads also try nodes containing replicated partitions,
     ;; but writes remain on master node.)))
     ;; This option requires ClientPolicy.requestProleReplicas to be enabled in order to function properly.
-    (is (= 30000 (.socketTimeout rp))) ;; 30 seconds default
-    (is (zero? (.totalTimeout rp))) ;; no time limit
-    (is (= 3000 (.timeoutDelay rp))) ;; no delay, connection closed on timeout
-    (is (= 2 (.maxRetries rp))) ;; initial attempt + 2 retries = 3 attempts
-    (is (zero? (.sleepBetweenRetries rp))) ;; do not sleep between retries
-    (is (false? (.sendKey rp))) ;; do not send the user defined key
-    (is (= ReadModeSC/SESSION (.readModeSC rp))))) ;; Ensures this client will only see an increasing sequence of record versions. Server only reads from master. This is the default..
+    ;; This option requires ClientPolicy.requestProleReplicas to be enabled in order to function properly.
+    (is (= 30000 (.socketTimeout rp)))                      ;; 30 seconds default
+    (is (= 1000 (.totalTimeout rp)))                        ;; total timeout of 1 second
+    (is (= 3000 (.timeoutDelay rp)))                        ;; no delay, connection closed on timeout
+    (is (= 2 (.maxRetries rp)))                             ;; initial attempt + 2 retries = 3 attempts
+    (is (zero? (.sleepBetweenRetries rp)))                  ;; do not sleep between retries
+    (is (false? (.sendKey rp)))                             ;; do not send the user defined key
+    (is (= ReadModeSC/SESSION (.readModeSC rp)))))          ;; Ensures this client will only see an increasing sequence of record versions. Server only reads from master. This is the default.
 
 (deftest configure-write-policy
   (let [c  (client/init-simple-aerospike-client
-             ["localhost"] "test"
+             *as-hosts* as-namespace
              {"writePolicyDefault" (policy/map->write-policy {"CommitLevel"        "COMMIT_MASTER"
                                                               "durableDelete"      true
                                                               "expiration"         1000
@@ -598,7 +639,6 @@
                                                               "respondAllOps"      true})})
 
         wp ^WritePolicy (.getWritePolicyDefault ^AerospikeClient (.-client ^SimpleAerospikeClient c))]
-    (is (= Priority/DEFAULT (.priority wp)))
     (is (= ReadModeAP/ONE (.readModeAP wp)))
     (is (true? (.durableDelete wp)))
     (is (= 1000 (.expiration wp)))
@@ -607,6 +647,34 @@
     (is (= RecordExistsAction/REPLACE_ONLY (.recordExistsAction wp)))
     (is (true? (.respondAllOps wp)))))
 
+(deftest configure-batch-write-policies
+  (let [expression                (Exp/build (Exp/ge (Exp/intBin "a") (Exp/intBin "b")))
+        c                         (client/init-simple-aerospike-client
+                                    *as-hosts* as-namespace
+                                    {"batchWritePolicyDefault"       (policy/map->batch-write-policy {"CommitLevel"        "COMMIT_MASTER"
+                                                                                                      "durableDelete"      true
+                                                                                                      "expiration"         1000
+                                                                                                      "generation"         7
+                                                                                                      "GenerationPolicy"   "EXPECT_GEN_GT"
+                                                                                                      "RecordExistsAction" "REPLACE_ONLY"
+                                                                                                      "filterExp"          expression})
+                                     "batchParentPolicyWriteDefault" (policy/map->batch-policy {"allowInline"          false
+                                                                                                "maxConcurrentThreads" 2
+                                                                                                "sendSetName"          true})})
+
+        batch-write-policy        ^BatchWritePolicy (.getBatchWritePolicyDefault ^AerospikeClient (.-client ^SimpleAerospikeClient c))
+        batch-parent-write-policy ^BatchPolicy (.getBatchParentPolicyWriteDefault ^AerospikeClient (.-client ^SimpleAerospikeClient c))]
+    (is (true? (.durableDelete batch-write-policy)))
+    (is (= 1000 (.expiration batch-write-policy)))
+    (is (= 7 (.generation batch-write-policy)))
+    (is (= GenerationPolicy/EXPECT_GEN_GT (.generationPolicy batch-write-policy)))
+    (is (= RecordExistsAction/REPLACE_ONLY (.recordExistsAction batch-write-policy)))
+    (is (= expression (.filterExp batch-write-policy)))
+    (is (= CommitLevel/COMMIT_MASTER (.commitLevel batch-write-policy)))
+
+    (is (false? (.allowInline batch-parent-write-policy)))
+    (is (= 2 (.maxConcurrentThreads batch-parent-write-policy)))
+    (is (true? (.sendSetName batch-parent-write-policy)))))
 
 (deftest set-entry
   (let [data        (rand-int 1000)
@@ -687,12 +755,11 @@
         @(pt/put-multiple *c* [k k2 k3] (repeat _set) [10 20 30] (repeat ttl) conf)
 
         @(pt/scan-set *c* aero-namespace _set {:callback callback})
-
+        (Thread/sleep 50)                                   ;wait for callback completion
         (let [res @(pt/get-batch *c* [{:index k :set _set}
                                       {:index k2 :set _set}
                                       {:index k3 :set _set}])]
-
-          (is (= (sort (mapv :payload res)) [11 21 31]))))
+          (is (= (mapv :payload res) [11 21 31]))))
       (delete-records))
 
     (testing "it can delete items during a scan"
@@ -725,3 +792,61 @@
         (is (= 1 (count @res))))
 
       (delete-records))))
+
+(deftest client-events-test
+  (let [success-spy               (spy/spy)
+        failure-spy               (spy/spy)
+        create-mock-client-events (fn [on-success-spy on-failure-spy]
+                                    (reify pt/ClientEvents
+                                      (on-success [_this op-name op-result index op-start-time]
+                                        (on-success-spy op-name op-result index op-start-time)
+                                        op-result)
+                                      (on-failure [_this op-name op-ex index op-start-time]
+                                        (on-failure-spy op-name op-ex index op-start-time)
+                                        op-ex)))
+        client-events             (create-mock-client-events success-spy failure-spy)
+        c                         (client/init-simple-aerospike-client *as-hosts* as-namespace {:client-events client-events})
+        index                     (random-key)
+        payload                   {"string" "hello"}]
+    (letfn [(nth-args-map [spy-instance ks n]
+              (let [args-map (zipmap ks (spy/nth-call spy-instance n))]
+                (assert number? (:op-start-time args-map))
+                (dissoc args-map :op-start-time)))
+            (success-nth-args-map [spy-instance n]
+              (nth-args-map spy-instance [:op-name :op-result :index :op-start-time] n))
+            (failure-nth-args-map [spy-instance n]
+              (nth-args-map spy-instance [:op-name :op-ex :index :op-start-time] n))]
+      (testing "client events is called with the right parameters using default client-events passed on client initialization"
+        @(pt/create c index _set payload TTL)
+        @(pt/get-single c index _set)
+        (is (thrown-with-msg? ExecutionException #"Generation error" @(pt/update c index _set "new-payload" 42 TTL)))
+
+        (is (= {:op-name   :write
+                :op-result true
+                :index     index} (success-nth-args-map success-spy 0)))
+        (is (= {:op-name   :read
+                :op-result payload
+                :index     index} (-> (success-nth-args-map success-spy 1)
+                                      (update :op-result :payload))))
+        (is (= {:op-name :write
+                :op-ex   3
+                :index   index} (-> (failure-nth-args-map failure-spy 0)
+                                    (update :op-ex #(.getResultCode %))))))
+      (testing "client events is called with the right parameters using custom client-events per op"
+        (let [custom-success-spy   (spy/spy)
+              custom-failure-spy   (spy/spy)
+              custom-client-events (create-mock-client-events custom-success-spy custom-failure-spy)]
+          @(pt/put c index _set payload TTL {:client-events [custom-client-events]})
+          @(pt/get-single c index _set {:client-events [custom-client-events]})
+          (is (thrown-with-msg? ExecutionException #"Generation error" @(pt/update c index _set "new-payload" 42 TTL {:client-events [custom-client-events]})))
+          (is (= {:op-name   :write
+                  :op-result true
+                  :index     index} (success-nth-args-map custom-success-spy 0)))
+          (is (= {:op-name   :read
+                  :op-result payload
+                  :index     index} (-> (success-nth-args-map custom-success-spy 1)
+                                        (update :op-result :payload))))
+          (is (= {:op-name :write
+                  :op-ex   3
+                  :index   index} (-> (failure-nth-args-map custom-failure-spy 0)
+                                      (update :op-ex #(.getResultCode %))))))))))