+++ update-docs.sh
@@ -10,7 +10,7 @@ git clean -fdx
 cd ..
 echo "regenerate docs..."
 
-lein with-profile docs codox
+lein with-profile +docs codox
 
 echo "commit changes"
 cd codox